refactoring

+++ src/main/java/by/estore/bean/Role.java
@@ -6,16 +6,16 @@ import java.util.Objects;
 public class Role implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Byte id;
     private String name;
 
     public Role() {}
 
-    public Long getId() {
+    public Byte getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Byte id) {
         this.id = id;
     }
